OnlinePavadzimes/requirements.txt
streamlit
reportlab
python-docx
beautifulsoup4
requests
num2words
Pillow
pandas
